+++ lib/constants.ts
@@ -2,8 +2,6 @@ const Url = process.env.NEXT_PUBLIC_BASE_API
   ? `${process.env.NEXT_PUBLIC_BASE_API}/`
   : "https://codebility-be.onrender.com/api/v1/production/"
 
-console.log(Url)
-
 export const API = {
   CODEVS: `${Url}users`,
   AUTH_GOOGLE_CODEVS: `${Url}user`,
